Notify RTImage on scaling change
The code only updates scalable icons on first initialization. That means
the icons do not resize if the user changes the scaling by settings or by
dragging to a different monitor.

Another edge case is if the user has a monitor with lower scaling that
is the initialization monitor but is not the main display monitor.
The scaling and icons are initialized for the init monitor, but the app is
displayed on the main display monitor.

This CL adds a global signal that is emitted when the scaling changes.
RTImage will refresh its cached image and request a redraw when this
signal is received.

Limitations:
The current infrastructure is not window-aware. If there are multiple
windows open, the scaling is applied globally which could lead to
discrepencies between windows.